pyproject.toml
[tool.poetry]
name = "fal"
version = "0.11.3a0"
description = "fal is an easy-to-use Serverless Python Framework"
authors = ["Features & Labels <hello@fal.ai>"]
readme = "README.md"
packages = [
    { include = "fal", from = "src" },
    { include = "openapi_fal_rest", from = "openapi-fal-rest" },
]

[tool.poetry.dependencies]
python = "^3.8"
isolate = {version = ">=0.12.3, <1.0", extras = ["build"]}
isolate-proto = "^0.2.0"
auth0-python = "^4.1.0"
requests = "^2.28.1"
grpcio = "^1.50.0"
dill = "0.3.7"
typing-extensions = "^4.7.1"
click = "^8.1.3"
structlog = "^22.3.0"
datadog-api-client = "2.12.0"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
grpc-interceptor = "^0.15.0"
colorama = "^0.4.6"
portalocker = "^2.7.0"
rich = "^13.3.2"
packaging = ">=21.3"
pathspec = "^0.11.1"
pydantic = "<2.0"

# serve=True dependencies
fastapi = "0.99.1"

# rest-api-client dependencies
httpx = ">=0.15.4,<0.25.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"
types-python-dateutil = "^2.8.0"

# For 3.9 and earlier, importlib-metadata's newer versions are included in the standard library.
importlib-metadata = { version = ">=4.4", python = "<3.10" }
boto3 = "^1.33.8"

[tool.poetry.group.dev.dependencies]
openapi-python-client = "^0.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# TODO: deprecated fal-serverless CLI
fal-serverless = "fal.cli:cli"
fal = "fal.cli:cli"
